Separate cell card component
